Implemented a new constructor for System.Lazy<T>.
Implemented a new constructor for System.Lazy<T> that can be used to simplify
and optimize code which may produce either an eagerly or lazily initialized
value. Consumers can handle both cases uniformly if the eagerly initialized
value is wrapped in a Lazy<T> instance. This currently requires creation of a
thunk to return the value. The new constructor avoids creating a thunk;
furthermore, the created Lazy<T> instance's IsValueCreated property will be true
immediately after the constructor returns.