Optimize implementations of Attribute.GetCustomAttribute() overloads.
Implemented internal methods on derived "runtime" classes such as
RuntimeFieldInfo and RuntimeType which allow for minimally-allocating
implementations of the Attribute.GetCustomAttribute() overloads;
in turn, these internal methods call into specific overloads of the
newly-defined CustomAttribute.GetCustomAttribute() method.
The CustomAttribute.GetCustomAttribute() overloads are currently just
shims calling the corresponding GetCustomAttributes() method; in order
to eliminate the array allocations associated with
GetCustomAttributes(), the GetCustomAttribute() overloads need to be
re-implemented to return the number of attribute instances and a single
attribute instance (when exactly one instance is present).